backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pandas==2.1.4
numpy==1.26.3
yfinance==0.2.33
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
scipy==1.11.4